Refactor content module
